Extract relational API and batch context
