+++ crates/truthdb-core/src/rel/mod.rs
@@ -8,10 +8,25 @@
 //! error numbers.
 
 mod aggregate;
+mod api;
+mod batch;
 pub mod collation;
+mod context;
 mod hash;
 mod plan;
 mod value;
+
+pub use api::{
+    BatchEmitter, BatchOutcome, Collector, DoneCommand, FATAL_SEVERITY, ResultColumn, RowSet,
+    RpcParam, StatementResult,
+};
+pub use batch::{execute_batch, execute_batch_streamed, execute_batch_with_params};
+pub use context::{Isolation, TxnContext};
+
+use api::{DeferredDone, DiscardEmitter, XACT_ABORT_SEVERITY};
+use batch::BatchRun;
+use batch::done_command;
+use context::TableVar;
 
 use truthdb_sql::ast::{
     AlterAction, AlterDatabase, AlterTable, CheckConstraint, ColumnDef, CreateFunction,
@@ -42,824 +57,15 @@ use crate::relstore::types::{ColumnType, Datum};
 use crate::relstore::version::ReadSnapshot;
 use crate::storage::{RowLocator, Storage, StorageError, StorageTxn, TxnScope};
 
-/// A declared table variable's in-memory contents: its column schema, the key
-/// columns of its declared PRIMARY KEY (for uniqueness enforcement), the
-/// per-column `DEFAULT` source text (parallel to the schema columns, re-parsed
-/// and evaluated per INSERT), and its rows. A row is a `Vec<Datum>` in schema
-/// order, exactly like a base-table row.
-#[derive(Clone)]
-struct TableVar {
-    schema: Schema,
-    key_columns: Vec<usize>,
-    defaults: Vec<Option<String>>,
-    rows: Vec<Vec<Datum>>,
-}
-
-/// truthdb-sql cannot depend on this crate, so it mirrors the default
-/// database id; the two constants must never drift.
-const _: () = assert!(
-    truthdb_sql::eval::DEFAULT_DATABASE_ID == crate::relstore::catalog::DEFAULT_DATABASE_ID
-);
-
-/// A session's current database id. A wrapper so `TxnContext::default()`
-/// lands in the DEFAULT database (id 1), never a nonexistent id 0 — one
-/// representation of "the default database", not two.
-#[derive(Clone, Copy, PartialEq, Eq)]
-struct CurrentDb(u32);
-
-impl Default for CurrentDb {
-    fn default() -> Self {
-        CurrentDb(crate::relstore::catalog::DEFAULT_DATABASE_ID)
-    }
-}
-
-/// Per-session transaction state carried across statements/batches. Lives in
-/// the session (engine thread); autocommit statements use `Default`.
-#[derive(Default)]
-pub struct TxnContext {
-    txn: Option<StorageTxn>,
-    /// `@@TRANCOUNT` — nested BEGINs increment; only the outermost COMMIT
-    /// actually commits.
-    trancount: u32,
-    /// Set when a statement failed inside the transaction (SQL Server
-    /// XACT_ABORT-style): only ROLLBACK is then allowed.
-    doomed: bool,
-    xact_abort: bool,
-    isolation: Isolation,
-    /// `SET NOCOUNT ON` — statements report no row count to the client:
-    /// TDS DONEs drop `DONE_COUNT`, and the native protocol's count envelope
-    /// becomes a bare done (the CLI then prints no "(n rows affected)" line,
-    /// exactly as sqlcmd goes quiet against SQL Server). Result rows and
-    /// `@@ROWCOUNT` are untouched.
-    nocount: bool,
-    /// Rows affected/returned by the previous statement — `@@ROWCOUNT`.
-    rowcount: i64,
-    /// The previous statement's error number, 0 on success — `@@ERROR`.
-    last_error: i32,
-    /// Names of the stored procedures currently executing (innermost last),
-    /// for `ERROR_PROCEDURE()`; empty in an ad-hoc batch.
-    proc_stack: Vec<String>,
-    /// The procedure executing when the LAST error was raised — captured at
-    /// the raise site, because by CATCH entry the procedure's frame has
-    /// already unwound off `proc_stack`.
-    error_procedure: Option<String>,
-    /// A procedure body's `RETURN [value]` status, stashed by the Return arm
-    /// for `EXEC @rc = name` to read after the body unwinds; 0 when the body
-    /// falls off the end.
-    proc_return: Option<i64>,
-    /// A scalar function body's `RETURN <expr>` value, coerced to the declared
-    /// return type and stashed by the Return arm for the caller to read. Only
-    /// set while a function body runs (see [`run_user_scalar_function`]).
-    func_return: Option<SqlValue>,
-    /// `SET SHOWPLAN_TEXT ON` — a SELECT returns its plan text, not results.
-    showplan_text: bool,
-    /// Declared batch variables (name without `@`, lowercased) to their type
-    /// and current value. Cleared at the start of each batch.
-    variables: std::collections::HashMap<String, (ColumnType, SqlValue)>,
-    /// Declared table variables (name without `@`, lowercased): in-memory
-    /// rowsets that live only on the session (never on Storage), so they survive
-    /// ROLLBACK and are cleared at batch end — SQL Server table-variable
-    /// semantics. Kept disjoint from `variables`; a name lives in exactly one.
-    table_variables: std::collections::HashMap<String, TableVar>,
-    /// Declared cursors (name lowercased). Batch-scoped (cleared at batch end).
-    cursors: std::collections::HashMap<String, CursorState>,
-    /// `@@FETCH_STATUS`: the result of the last cursor FETCH (0 / -1 / -2).
-    fetch_status: i32,
-    /// Connection identity for session intrinsics (`DB_NAME()`,
-    /// `SUSER_SNAME()`, `@@SPID`), set once when the session opens.
-    database: String,
-    /// Per-batch snapshot of every database `(id, canonical name)` — read by
-    /// `DB_ID(name)`/`DB_NAME(id)`. Refreshed with the security context at
-    /// batch start (a same-batch CREATE DATABASE is visible next batch, a
-    /// documented staleness matching the per-batch security refresh).
-    databases_snapshot: Vec<(u32, String)>,
-    /// The session's current database id — the namespace every unqualified
-    /// object name resolves in. Sessions run in the default database until
-    /// `USE` learns to switch it (the multi-database plan's A2 slice).
-    database_id: CurrentDb,
-    login: String,
-    spid: i32,
-    /// The session's database user name (`USER_NAME()`), resolved from the login
-    /// when the session opens (`dbo` for a sysadmin, else the mapped user, else
-    /// the login name). Distinct from `login` (the server principal).
-    user: String,
-    /// The login's server principal_id and the user's database principal_id —
-    /// the keys the membership intrinsics resolve against. Set at session open.
-    login_sid: u32,
-    user_sid: u32,
-    /// The session's effective SERVER-role names (lowercased) — the transitive
-    /// closure of the login's roles (today: `sysadmin`). Read by
-    /// `IS_SRVROLEMEMBER`. Kept separate from database roles so the two
-    /// namespaces do not cross-answer.
-    session_server_roles: std::collections::HashSet<String>,
-    /// The session's effective DATABASE-role names (lowercased) — the transitive
-    /// closure of the database user's roles. Read by `IS_ROLEMEMBER`. Both sets
-    /// are refreshed at batch start from the membership cache, so a security DDL
-    /// is seen by the next batch (SQL Server's per-batch permission caching).
-    session_db_roles: std::collections::HashSet<String>,
-    /// Object-permission enforcement subject (bypass flag + the grant-matching
-    /// principal id set), refreshed at batch start alongside the role sets.
-    security: truthdb_sql::eval::SecurityContext,
-    /// The last identity value inserted in this session (SQL Server scope),
-    /// surfaced as `SCOPE_IDENTITY()`. Persists across statements until the next
-    /// identity INSERT; unaffected by non-identity inserts.
-    scope_identity: Option<i64>,
-    /// Named savepoints in the current transaction (`SAVE TRANSACTION <name>`,
-    /// lowercased) → the point to which `ROLLBACK TRANSACTION <name>` returns.
-    /// Cleared when the transaction ends.
-    savepoints: std::collections::HashMap<String, crate::relstore::ctx::Savepoint>,
-    /// Errors caught by the currently-executing `CATCH` blocks (a stack, so
-    /// nested `TRY`/`CATCH` restore the outer error on exit). `ERROR_*()` read
-    /// the top; empty outside any `CATCH` block.
-    error_stack: Vec<truthdb_sql::eval::ErrorInfo>,
-    /// The SNAPSHOT-isolation transaction's read view, captured (and
-    /// registered against pruning) at its first data access and released
-    /// when the transaction ends — commit, rollback, reap, or disconnect.
-    txn_snapshot: Option<ReadSnapshot>,
-    /// Set when the idle reaper rolled this session's transaction back. The
-    /// session's next batch fails with 1205 and clears it, so a client that
-    /// comes back believing it is still in a transaction is told the
-    /// transaction is gone — rather than silently autocommitting statements it
-    /// means to be transactional, and only discovering it at a COMMIT that
-    /// errors 3902 long after the writes became durable.
-    reaped: bool,
-}
-
-/// Session isolation level (defaults to READ COMMITTED, like SQL Server).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum Isolation {
-    ReadUncommitted,
-    #[default]
-    ReadCommitted,
-    RepeatableRead,
-    Serializable,
-    /// Transaction-scoped versioned reads (Stage 13): one snapshot at the
-    /// transaction's first data access, reused by every statement in it.
-    Snapshot,
-}
-
-impl TxnContext {
-    fn in_txn(&self) -> bool {
-        self.txn.is_some()
-    }
-
-    fn eval_context(&self) -> EvalContext {
-        EvalContext {
-            database_id: self.database_id.0,
-            databases: self.databases_snapshot.clone(),
-            trancount: self.trancount as i32,
-            variables: self
-                .variables
-                .iter()
-                .map(|(name, (_, value))| (name.clone(), value.clone()))
-                .collect(),
-            database: self.database.clone(),
-            login: self.login.clone(),
-            user: self.user.clone(),
-            server_roles: self.session_server_roles.clone(),
-            db_roles: self.session_db_roles.clone(),
-            security: self.security.clone(),
-            spid: self.spid,
-            rowcount: self.rowcount,
-            scope_identity: self.scope_identity,
-            error: self.error_stack.last().cloned(),
-            xact_state: self.xact_state(),
-            last_error: self.last_error,
-            nestlevel: EXEC_DEPTH.with(|d| d.get()) as i32,
-            updated_columns: current_trigger_updated_columns(),
-            fetch_status: self.fetch_status,
-        }
-    }
-
-    /// `XACT_STATE()`: 0 with no open transaction, -1 when the open transaction
-    /// is doomed (uncommittable), else 1.
-    fn xact_state(&self) -> i8 {
-        if !self.in_txn() {
-            0
-        } else if self.doomed {
-            -1
-        } else {
-            1
-        }
-    }
-
-    /// Enters a `CATCH` block: records the caught error so `ERROR_*()` resolve
-    /// to it (pushed, so nested `TRY`/`CATCH` restore the outer error on exit).
-    /// Records the error context every failed statement leaves behind:
-    /// `@@ERROR` and the raising procedure (for `ERROR_PROCEDURE()`). Called
-    /// at the RAISE site — the only place the procedure frame is still live.
-    fn record_error(&mut self, number: i32) {
-        self.last_error = number;
-        self.error_procedure = self.proc_stack.last().cloned();
-    }
-
-    fn push_error(&mut self, error: &SqlError) {
-        self.error_stack.push(truthdb_sql::eval::ErrorInfo {
-            number: error.number,
-            message: error.message.clone(),
-            severity: error.level,
-            state: error.state,
-            procedure: self.error_procedure.clone(),
-        });
-    }
-
-    /// Leaves a `CATCH` block, restoring the enclosing error context (if any).
-    fn pop_error(&mut self) {
-        self.error_stack.pop();
-    }
-
-    /// Records the connection identity used by session intrinsics. Called once
-    /// when the session opens. `session_roles` is filled separately, per batch,
-    /// from the membership cache (see [`Self::refresh_session_roles`]).
-    #[allow(clippy::too_many_arguments)]
-    pub fn set_session_identity(
-        &mut self,
-        database: String,
-        database_id: u32,
-        login: String,
-        spid: i32,
-        user: String,
-        login_sid: u32,
-        user_sid: u32,
-    ) {
-        self.database = database;
-        self.database_id = CurrentDb(database_id);
-        self.login = login;
-        self.spid = spid;
-        self.user = user;
-        self.login_sid = login_sid;
-        self.user_sid = user_sid;
-    }
-
-    /// The session's current database id — the namespace unqualified names
-    /// resolve in.
-    pub(crate) fn database_id(&self) -> u32 {
-        self.database_id.0
-    }
-
-    /// Switches the session's current database (`USE`): the canonical name
-    /// and its id move together — they are one decision, never two.
-    pub(crate) fn set_current_database(&mut self, name: String, db_id: u32) {
-        self.database = name;
-        self.database_id = CurrentDb(db_id);
-    }
-
-    /// Refreshes the session's effective role NAMES from the membership cache.
-    /// The login's roles are SERVER roles (`sysadmin`); the database user's roles
-    /// are DATABASE roles — kept in separate sets so `IS_SRVROLEMEMBER` and
-    /// `IS_ROLEMEMBER` never answer for the other's namespace. Called at batch
-    /// start; a security DDL is therefore visible to the next batch.
-    pub fn refresh_session_roles(&mut self, storage: &Storage) {
-        let server_role_ids = storage.effective_roles(self.login_sid);
-        let db_role_ids = storage.effective_roles(self.user_sid);
-        let names = |ids: &std::collections::HashSet<u32>| -> std::collections::HashSet<String> {
-            ids.iter()
-                .filter_map(|&id| storage.principal_name(id))
-                .map(|name| name.to_ascii_lowercase())
-                .collect()
-        };
-        self.session_server_roles = names(&server_role_ids);
-        self.session_db_roles = names(&db_role_ids);
-        self.databases_snapshot = storage.rel_databases();
-
-        // The object-permission subject. A trusted/internal connection
-        // (login_sid 0 — the native protocol and in-process tests), a sysadmin,
-        // or dbo/db_owner bypasses every object-permission check (owns or
-        // controls the database). Otherwise a GRANT/DENY matches the database
-        // user, its effective roles, or `public`.
-        use crate::storage::{DB_OWNER_ID, DBO_ID, PUBLIC_ID, SYSADMIN_ID};
-        let bypass = self.login_sid == 0
-            || server_role_ids.contains(&SYSADMIN_ID)
-            || self.user_sid == DBO_ID
-            || db_role_ids.contains(&DB_OWNER_ID);
-        let mut principals = db_role_ids;
-        principals.insert(self.user_sid);
-        principals.insert(PUBLIC_ID);
-        self.security = truthdb_sql::eval::SecurityContext { bypass, principals };
-    }
-
-    /// Clears batch-scoped variables (called at the start of each batch).
-    pub fn clear_variables(&mut self) {
-        self.variables.clear();
-        self.table_variables.clear();
-        self.cursors.clear();
-        self.fetch_status = 0;
-    }
-
-    /// The final value and type of a batch variable, as a `Datum`, for the
-    /// RPC-by-name response tail: after the synthesized `EXEC` batch completes
-    /// the session reads the OUTPUT parameters (copied back into caller-scope
-    /// variables) and the seeded return-status variable back off the context.
-    /// `name` may carry a leading `@`; lookup is case-insensitive, matching how
-    /// variables are keyed.
-    pub fn variable_datum(&self, name: &str) -> Option<(ColumnType, Datum)> {
-        let key = name.trim_start_matches('@').to_ascii_lowercase();
-        let (column_type, value) = self.variables.get(&key)?;
-        let datum = value::sql_to_datum(value, column_type, &key).ok()?;
-        Some((*column_type, datum))
-    }
-
-    /// True if a transaction is open (used by the session to decide whether a
-    /// disconnect must roll back).
-    pub fn has_open_transaction(&self) -> bool {
-        self.txn.is_some()
-    }
-
-    /// Whether an explicit transaction is open (`@@TRANCOUNT > 0`) — what a
-    /// reply's DONE stamps as `DONE_INXACT`.
-    pub fn in_transaction(&self) -> bool {
-        self.trancount > 0
-    }
-
-    /// The session's current isolation level (drives which locks reads take).
-    pub fn isolation(&self) -> Isolation {
-        self.isolation
-    }
-
-    /// Rolls back and discards any open transaction, resetting every piece of
-    /// transaction-scoped state.
-    ///
-    /// `savepoints` must be cleared here, not merely on the paths that discard
-    /// the context afterwards: a savepoint holds the *undo-log offset* of the
-    /// transaction that recorded it, so one surviving into the session's next
-    /// transaction would let `ROLLBACK TRANSACTION <name>` find a stale entry
-    /// instead of erroring 3908 — and hand a dead transaction's offset to
-    /// `rel_rollback_to`, which either truncates the new transaction's undo log
-    /// (silently discarding committed work) or panics on `split_off`.
-    pub fn abort(&mut self, storage: &Storage) {
-        if let Some(txn) = self.txn.take() {
-            let _ = storage.rel_rollback(txn);
-        }
-        self.release_txn_snapshot(storage);
-        self.trancount = 0;
-        self.doomed = false;
-        self.savepoints.clear();
-    }
-
-    /// Releases the SNAPSHOT transaction's registered read view, if any —
-    /// called on every transaction-ending path (a leaked registration pins
-    /// the version store's prune watermark forever).
-    fn release_txn_snapshot(&mut self, storage: &Storage) {
-        if let Some(snap) = self.txn_snapshot.take() {
-            storage.release_read_snapshot(snap.seq);
-        }
-    }
-
-    /// Rolls back this session's transaction because it sat idle too long, and
-    /// records that it happened so the session's next batch can say so.
-    pub fn abort_idle(&mut self, storage: &Storage) {
-        self.abort(storage);
-        self.reaped = true;
-    }
-
-    /// Takes the "your transaction was reaped" flag, if set.
-    fn take_reaped(&mut self) -> bool {
-        std::mem::take(&mut self.reaped)
-    }
-}
-
-/// Result of one executed statement.
-#[derive(Debug, Clone, PartialEq)]
-pub enum StatementResult {
-    Rows(RowSet),
-    RowsAffected(u64),
-    /// DDL and other statements with no rowset and no count.
-    Done,
-}
-
-/// A result column: its name and resolved SQL type (drives TDS
-/// COLMETADATA and display rendering alike).
-#[derive(Debug, Clone, PartialEq)]
-pub struct ResultColumn {
-    pub name: String,
-    pub column_type: ColumnType,
-}
-
-/// A typed result set: column metadata plus rows of typed [`Datum`]s.
-#[derive(Debug, Clone, PartialEq)]
-pub struct RowSet {
-    pub columns: Vec<ResultColumn>,
-    pub rows: Vec<Vec<Datum>>,
-}
-
-/// Error severity at or above which a statement failure dooms the whole
-/// transaction even under `SET XACT_ABORT OFF` (SQL Server treats severity ≥ 17
-/// as resource/batch-level, versus 11–16 statement-level). Constraint violations
-/// (2627/2601/515/547, severity 14–16) stay below it, so they roll back only the
-/// failing statement and the transaction survives.
-const XACT_ABORT_SEVERITY: u8 = 17;
-
-/// Error severity at or above which the error is fatal to the CONNECTION
-/// (SQL Server severity >= 20): it bypasses every `TRY`, dooms the
-/// transaction, and the protocol layers close the stream after delivering
-/// it. Only RAISERROR ... WITH LOG can currently produce one.
-pub const FATAL_SEVERITY: u8 = 20;
-
-/// A batch's outcome: the results of the statements that ran, plus the error
-/// that stopped the batch (if any). Statements before an error have already
-/// committed (each statement is autocommit in Stage 3), so their results are
-/// preserved rather than discarded.
-pub struct BatchOutcome {
-    pub results: Vec<StatementResult>,
-    pub error: Option<SqlError>,
-}
-
-/// One `sp_executesql` parameter: its `@name` (as it appears in the RPC
-/// stream), declared type, and decoded value. Passed by the TDS layer to
-/// [`execute_batch_with_params`], which seeds them as batch variables the
-/// statement text can read by name.
-#[derive(Debug, Clone)]
-pub struct RpcParam {
-    pub name: String,
-    pub column_type: ColumnType,
-    pub value: Datum,
-}
-
-/// Parses and executes a SQL batch. A parse error yields an empty batch with
-/// the error; a runtime error stops the batch but keeps earlier results.
-pub fn execute_batch(storage: &Storage, sql: &str, txn_ctx: &mut TxnContext) -> BatchOutcome {
-    execute_batch_with_params(storage, sql, txn_ctx, &[])
-}
-
-/// Like [`execute_batch`], but seeds `params` as batch variables before running
-/// the statement text — the `sp_executesql` path. Parameters are injected as
-/// already-typed values, never re-rendered into the SQL text, so a parameter
-/// value can never alter the statement's structure (no injection surface).
-pub fn execute_batch_with_params(
-    storage: &Storage,
-    sql: &str,
-    txn_ctx: &mut TxnContext,
-    params: &[RpcParam],
-) -> BatchOutcome {
-    let mut collector = Collector::default();
-    let error = execute_batch_streamed(storage, sql, txn_ctx, params, &mut collector);
-    collector.into_outcome(error)
-}
-
-/// Like [`execute_batch_with_params`], but each statement's result leaves
-/// through `emitter` as the statement produces it instead of accumulating into
-/// a [`BatchOutcome`]: a result set opens with its columns, its rows follow (in
-/// chunks read straight off the scan for the streamed shape), and each
-/// statement's DONE carries the transaction state *after that statement* —
-/// which is what TDS `DONE_INXACT` means per statement. The returned error is
-/// the batch's terminal error (what `BatchOutcome::error` carried), reported by
-/// the caller after the statement events.
-///
-/// Durability keeps its ordering: nothing the client can rely on — a DONE
-/// acknowledging a commit, or rows carrying commit-derived state such as a
-/// reserved identity value — is emitted before the commit behind it is
-/// fsync-durable. DONEs queue in the run and flush before the next result set
-/// opens and at the end of the batch; both points fsync first when any
-/// statement since the last one may have committed (the same kind-based test
-/// as before), so a batch of writes with nothing to stream between them still
-/// costs one fsync.
-pub fn execute_batch_streamed(
-    storage: &Storage,
-    sql: &str,
-    txn_ctx: &mut TxnContext,
-    params: &[RpcParam],
-    emitter: &mut dyn BatchEmitter,
-) -> Option<SqlError> {
-    // A transaction reaped for idleness is reported to the session's next batch
-    // (once). 1205 is the code this engine already uses for a server-initiated
-    // transaction abort (the parked deadlock victim), and every driver treats it
-    // as "the transaction is gone, retry it" — which is exactly the right
-    // recovery here.
-    if txn_ctx.take_reaped() {
-        return Some(SqlError::new(
-            1205,
-            13,
-            51,
-            "The transaction was rolled back because the session was idle for too long. Rerun the transaction.",
-        ));
-    }
-    // Variables are batch-scoped: each batch starts with none.
-    txn_ctx.clear_variables();
-    // Refresh the session's effective role set from the membership cache, so a
-    // security DDL committed since the last batch is reflected in this one's
-    // IS_ROLEMEMBER/IS_SRVROLEMEMBER (SQL Server's per-batch permission caching).
-    txn_ctx.refresh_session_roles(storage);
-    for param in params {
-        // The lexer keys `@p1` as `p1` (leading `@` stripped, lowercased); the
-        // RPC name arrives as `@p1`, so normalise it the same way to match.
-        let key = param.name.trim_start_matches('@').to_ascii_lowercase();
-        let value = value::datum_to_sql(&param.value, &param.column_type);
-        txn_ctx.variables.insert(key, (param.column_type, value));
-    }
-    let statements = match truthdb_sql::parse(sql) {
-        Ok(statements) => statements,
-        Err(error) => return Some(error),
-    };
-    let mut run = BatchRun {
-        emitter,
-        deferred: Vec::new(),
-        rowset_open: false,
-        durability_failed: false,
-        committed: false,
-        last_error: None,
-        function_return_type: None,
-    };
-    // `run_block` returns Err only when the batch must terminate (a cancel, or a
-    // dooming/uncaught error outside any TRY); a non-dooming error under
-    // `XACT_ABORT OFF` is recorded in `run.last_error` and the batch continues.
-    let terminating = run_block(storage, &statements, txn_ctx, &mut run, false)
-        .and_then(end_of_scope)
-        .err();
-    // The batch-end durability point, and the DONEs it was holding back. A
-    // durability failure outranks any statement error: a lost commit is more
-    // severe than an error the client asked about, and a benign continued error
-    // must not mask one.
-    let durability = run.finish(storage);
-    durability.or(terminating).or(run.last_error)
-}
-
-/// Receives a batch's results as the executor produces them. The session layer
-/// forwards each call as a `BatchEvent` onto the reply channel; buffered
-/// callers (the native command path, the SLT runner, tests) use [`Collector`]
-/// to reassemble a [`BatchOutcome`].
-pub trait BatchEmitter {
-    /// Opens a result set: its column metadata.
-    fn columns(&mut self, columns: Vec<ResultColumn>);
-    /// A chunk of rows for the open result set.
-    fn rows(&mut self, rows: Vec<Vec<Datum>>);
-    /// Ends one statement: its row count / rows-affected (`None` for DDL),
-    /// the transaction state after it ran, and its command class (the DONE's
-    /// `CurCmd` on the wire).
-    fn statement_done(&mut self, count: Option<u64>, in_transaction: bool, command: DoneCommand);
-    /// Ends a statement that failed after its result set had begun streaming:
-    /// the set is closed so the stream stays framed for the statements that
-    /// follow. The error itself is reported separately — at the end of the
-    /// batch for a continued error, or not at all for one a `CATCH` handled.
-    fn statement_aborted(&mut self, in_transaction: bool);
-
-    /// The session's database context was (re-)established (`USE`): TDS
-    /// renders the ENVCHANGE + 5701 INFO SSMS expects. Emitters that have no
-    /// wire (the collecting native path, tests) ignore it.
-    fn database_context(&mut self, _database: &str) {}
-
-    /// An informational message (RAISERROR severity <= 10): TDS renders an
-    /// INFO token in-stream, not an error. Emitters with no wire ignore it.
-    fn info(&mut self, _error: &SqlError) {}
-}
-
-/// A [`BatchEmitter`] that drops everything: a scalar function body produces no
-/// result sets (a data-returning SELECT is rejected at CREATE, 444), so its
-/// per-statement DONE events have nowhere to go.
-struct DiscardEmitter;
-
-impl BatchEmitter for DiscardEmitter {
-    fn columns(&mut self, _columns: Vec<ResultColumn>) {}
-    fn rows(&mut self, _rows: Vec<Vec<Datum>>) {}
-    fn statement_done(
-        &mut self,
-        _count: Option<u64>,
-        _in_transaction: bool,
-        _command: DoneCommand,
-    ) {
-    }
-    fn statement_aborted(&mut self, _in_transaction: bool) {}
-}
-
-/// Reassembles emitted results into the whole-batch [`BatchOutcome`] for the
-/// callers that want everything at once.
-#[derive(Default)]
-pub struct Collector {
-    results: Vec<StatementResult>,
-    /// The result set currently streaming, if a statement opened one.
-    open: Option<RowSet>,
-}
-
-impl Collector {
-    /// The collected outcome. A result set still open belongs to a statement
-    /// that failed after its rows started streaming; a failed statement
-    /// contributes no result, so it is dropped.
-    pub fn into_outcome(self, error: Option<SqlError>) -> BatchOutcome {
-        BatchOutcome {
-            results: self.results,
-            error,
-        }
-    }
-}
-
-impl BatchEmitter for Collector {
-    fn columns(&mut self, columns: Vec<ResultColumn>) {
-        self.open = Some(RowSet {
-            columns,
-            rows: Vec::new(),
-        });
-    }
-
-    fn rows(&mut self, mut rows: Vec<Vec<Datum>>) {
-        if let Some(rowset) = self.open.as_mut() {
-            rowset.rows.append(&mut rows);
-        }
-    }
-
-    fn statement_done(&mut self, count: Option<u64>, _in_transaction: bool, _command: DoneCommand) {
-        self.results.push(match self.open.take() {
-            Some(rowset) => StatementResult::Rows(rowset),
-            None => match count {
-                Some(n) => StatementResult::RowsAffected(n),
-                None => StatementResult::Done,
-            },
-        });
-    }
-
-    fn statement_aborted(&mut self, _in_transaction: bool) {
-        self.open = None;
-    }
-}
-
-/// The mutable accumulator threaded through [`run_block`] across a batch and
-/// its nested `TRY`/`CATCH` blocks: the emitter results leave through, the
-/// DONEs held back for durability, and the batch's error state.
-struct BatchRun<'a> {
-    emitter: &'a mut dyn BatchEmitter,
-    /// Finished statements' DONEs, held back until the next durability point
-    /// (the next result set opening, or the end of the batch) so a DONE that
-    /// acknowledges a commit is never emitted before that commit is durable.
-    deferred: Vec<DeferredDone>,
-    /// A result set's columns have been emitted but its statement has not
-    /// finished — the state [`BatchRun::abort_open_rowset`] closes on failure.
-    rowset_open: bool,
-    /// A durability (fsync) failure wedged the store. The error terminates the
-    /// batch and is never catchable: the old batch-end fsync ran past every
-    /// TRY, and a CATCH must not be able to swallow a lost commit.
-    durability_failed: bool,
-    /// Whether any executed statement may have made a durable commit: group
-    /// commit defers the WAL fsync, so the end of the batch fsyncs once if so.
-    committed: bool,
-    /// The last non-dooming statement error under `SET XACT_ABORT OFF` (outside
-    /// any TRY) — the batch continues past it (SQL Server default) and it is
-    /// reported alongside the results rather than terminating the batch.
-    last_error: Option<SqlError>,
-    /// Set while running a scalar function body: the declared return type. The
-    /// `RETURN <expr>` arm then evaluates its value, coerces it to this type,
-    /// and stashes it in `TxnContext::func_return` (rather than the procedure
-    /// int-status path).
-    function_return_type: Option<ColumnType>,
-}
-
-/// A statement's DONE, parked until the next durability point.
-struct DeferredDone {
-    count: Option<u64>,
-    in_transaction: bool,
-    command: DoneCommand,
-}
-
-/// The command class a statement's DONE reports in its `CurCmd` field.
-/// mssql-jdbc discards a DONE's row count unless `CurCmd` names a DML
-/// command, so `executeUpdate` returns -1 against a server that leaves it
-/// zero (pytds and go-mssqldb ignore the field).
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum DoneCommand {
-    Select,
-    Insert,
-    Update,
-    Delete,
-    /// DDL, SET, transaction control — anything whose count nobody reads.
-    Other,
-}
-
-/// The `CurCmd` class of a statement.
-fn done_command(statement: &Statement) -> DoneCommand {
-    match statement {
-        Statement::Select(_) => DoneCommand::Select,
-        Statement::Insert(_) => DoneCommand::Insert,
-        Statement::Update(_) => DoneCommand::Update,
-        Statement::Delete(_) => DoneCommand::Delete,
-        _ => DoneCommand::Other,
-    }
-}
-
-impl BatchRun<'_> {
-    /// Opens a result set. [`run_block`] flushed the deferred DONEs before any
-    /// statement that can produce one, so statement order on the stream holds.
-    fn open_rowset(&mut self, columns: Vec<ResultColumn>) {
-        debug_assert!(
-            self.deferred.is_empty(),
-            "a result set opened over deferred DONEs"
-        );
-        self.rowset_open = true;
-        self.emitter.columns(columns);
-    }
-
-    /// Emits a chunk of rows for the open result set.
-    fn rows(&mut self, rows: Vec<Vec<Datum>>) {
-        if !rows.is_empty() {
-            self.emitter.rows(rows);
-        }
-    }
-
-    /// Forwards a database-context change (`USE`) to the emitter, ahead of
-    /// the statement's (deferred) DONE.
-    fn database_context(&mut self, database: &str) {
-        self.emitter.database_context(database);
-    }
-
-    /// Emits an informational message (RAISERROR severity <= 10). `run_block`
-    /// flushed the deferred DONEs before the statement, so stream order holds.
-    fn info(&mut self, error: SqlError) {
-        debug_assert!(
-            self.deferred.is_empty(),
-            "an INFO message over deferred DONEs"
-        );
-        self.emitter.info(&error);
-    }
-
-    /// Ends a statement. Its DONE is deferred to the next durability point.
-    fn done(&mut self, count: Option<u64>, in_transaction: bool, command: DoneCommand) {
-        self.rowset_open = false;
-        self.deferred.push(DeferredDone {
-            count,
-            in_transaction,
-            command,
-        });
-    }
-
-    /// Closes the open result set of a statement that failed after its columns
-    /// (and possibly rows) were already emitted, so the stream stays framed for
-    /// the statements that follow (a caught or continued error). No-op when
-    /// nothing is open — a statement that failed before emitting anything
-    /// leaves no trace, as before.
-    fn abort_open_rowset(&mut self, in_transaction: bool) {
-        if self.rowset_open {
-            self.rowset_open = false;
-            self.emitter.statement_aborted(in_transaction);
-        }
-    }
-
-    /// Emits the deferred DONEs, fsyncing first when any statement since the
-    /// last durability point may have committed. The gate is the same
-    /// kind-based `committed` flag the batch-end fsync uses — not "does some
-    /// DONE acknowledge a commit" — because commit-derived state escapes
-    /// through the *rows* of whatever result set opens next, not only through
-    /// DONEs: an identity value reserved by an in-transaction INSERT (a
-    /// mini-commit) is readable one statement later via `SELECT
-    /// SCOPE_IDENTITY()`, and a value the client has seen must never be
-    /// reissued after a crash. On a durability failure the DONEs the batch
-    /// can no longer stand behind are dropped and the batch terminates (see
-    /// [`Self::make_durable`]).
-    fn flush(&mut self, storage: &Storage) -> Result<(), SqlError> {
-        if self.committed {
-            self.committed = false;
-            if let Some(error) = self.make_durable(storage) {
-                return Err(error);
-            }
-        }
-        for done in self.deferred.drain(..) {
-            self.emitter
-                .statement_done(done.count, done.in_transaction, done.command);
-        }
-        Ok(())
-    }
-
-    /// The end of the batch: one fsync if any statement may have committed
-    /// since the last durability point — by kind, not transaction state, so a
-    /// hidden mini-commit (an identity reservation, even inside an open
-    /// transaction or under a statement that then failed) is never missed —
-    /// then the remaining DONEs. Returns the durability error, if any.
-    fn finish(&mut self, storage: &Storage) -> Option<SqlError> {
-        if self.committed {
-            self.committed = false;
-            if let Some(error) = self.make_durable(storage) {
-                return Some(error);
-            }
-        }
-        for done in self.deferred.drain(..) {
-            self.emitter
-                .statement_done(done.count, done.in_transaction, done.command);
-        }
-        None
-    }
-
-    /// Blocks until the batch's commit records are fsync-durable (group
-    /// commit). A durability failure wedges the store — the in-memory state is
-    /// now ahead of the log, so no further op may serve it — and drops the
-    /// deferred DONEs, which would otherwise acknowledge commits a restart is
-    /// about to undo.
-    fn make_durable(&mut self, storage: &Storage) -> Option<SqlError> {
-        match storage.ensure_durable(storage.wal_tail()) {
-            Ok(()) => None,
-            Err(err) => {
-                storage.wedge();
-                self.deferred.clear();
-                self.durability_failed = true;
-                Some(map_storage_err(err, ""))
-            }
-        }
-    }
-}
-
+/// The inner SQL text of an `EXEC sp_executesql N'...'` whose statement
+/// argument is a string LITERAL — the analyzable case. `None` for any other
+/// procedure or a non-literal statement argument.
 /// Runs a statement list, recursing into `TRY`/`CATCH`. `in_try` is true while
 /// executing inside a `TRY` block, where a statement error transfers control to
 /// the matching `CATCH` (returned as `Err`) instead of applying the normal
 /// batch policy. Returns `Err` when the enclosing context must stop: a cancel,
 /// an error that propagates to a `CATCH`, or a dooming/terminating error at the
 /// top level.
-/// The inner SQL text of an `EXEC sp_executesql N'...'` whose statement
-/// argument is a string LITERAL — the analyzable case. `None` for any other
-/// procedure or a non-literal statement argument.
 fn exec_literal_sql(exec: &ExecStatement) -> Option<String> {
     if !strip_schema(&exec.proc.value).eq_ignore_ascii_case("sp_executesql") {
         return None;
